Hide URL bar on small screens.
That is necessary for me when I'm working with a splitted screen. Maybe there are others with this problem also.

Since it will not be used in mobile devices, then I guess it is safe to imagine the user will be able to expand the screen, input the URL, then reduce the screen again.

I think anything is better than not being able to click the "Execute Query" button.